refac to make me less sad
