+++ web-resource-uploader/Webresource.Uploader/Webresource.Uploader/Enum/WebresourceState.cs
@@ -1,0 +1,10 @@
+﻿namespace Webresource.Uploader.Enum
+{
+    public enum WebresourceState
+    {
+        New,
+        Draft,
+        Saved,
+        None
+    }
+}
