+++ web-resource-uploader/Webresource.Uploader/Webresource.Uploader/Enum/WebresourceType.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Webresource.Uploader.Enum
+{
+    public enum WebresourceType
+    {
+        WebPage = 1,
+        Css = 2,
+        Script = 3,
+        Data = 4,
+        Png = 5,
+        Jpg = 6,
+        Gif = 7,
+        Silverlight = 8,
+        Xsl = 9,
+        Ico = 10,
+        Vector = 11,
+        Resx = 12,
+        Auto = 99
+    }
+}
